Seperate Utils.ts and types.ts from Dashboard
